Moved testing functions into their own file
