+++ Core/Inc/main.h
@@ -31,7 +31,7 @@ extern "C" {
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
-#include <MPU6050.h>
+
 /* USER CODE END Includes */
 
 /* Exported types ------------------------------------------------------------*/
